moved Carousel to seperate component
